% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Add one to any value}
\usage{
add_one(value)
}
\arguments{
\item{value}{A numeric value}
}
\value{
Numeric. value + 1
}
\description{
Add one to any value
}
\examples{
add_one(12)
add_one(56)
}
